# dakunlun/docs/swagger.yaml
definitions:
...
  entity.Skill:
    properties:
      id:
        type: integer
      level:
        type: integer
    type: object
  entity.Skills:
    additionalProperties:
      $ref: '#/definitions/entity.Skill'
    type: object
  entity.Skin:
    properties:
      active:
        type: boolean
      fightingCapacity:
        type: integer
      val:
        type: integer
    type: object
  entity.SkinMap:
    additionalProperties:
      $ref: '#/definitions/entity.Skin'
    type: object
...
  msg.Explore:
    properties:
      exploreID:
        description: 配置表ID
        type: integer
      heroIDs:
        description: 武将自增ID列表
        items:
          type: integer
        type: array
      id:
        description: 自增ID
        type: integer
      mul:
        description: 倍数 1代表100%
        type: number
      startTime:
        description: 开始时间
        type: integer
    type: object
  msg.ExploreListResponse:
    properties:
      explores:
        items:
          $ref: '#/definitions/msg.Explore'
        type: array
      heros:
        items:
          $ref: '#/definitions/msg.Hero'
        type: array
    type: object
  msg.ExploreReceiveResponse:
    properties:
      rewards:
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
    type: object
  msg.ExploreSetResponse:
    properties:
      explore:
        $ref: '#/definitions/msg.Explore'
      rewards:
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
    type: object
...
  msg.GmCommonResponse:
    type: object
...
  msg.Hero:
    properties:
      attackFreq:
        type: integer
      attackTrans:
        type: integer
      defendTrans:
        type: integer
      evolveTimes:
        type: integer
      exploreID:
        type: integer
      fightingCapacity:
        type: integer
      heroID:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      race:
        type: integer
      sex:
        type: integer
      skills:
        $ref: '#/definitions/entity.Skills'
      skinID:
        type: integer
      skins:
        $ref: '#/definitions/entity.SkinMap'
      type:
        type: integer
    type: object
...
  msg.Reward:
    properties:
      mainType:
        type: integer
      subType:
        type: integer
      val:
        type: integer
    type: object
...
host: 8.140.113.88:8080
info:
  contact: {}
  description: 接口文档
  title: dakunlun
  version: "1.0"
paths:
...
  /api/explore/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.ExploreListResponse'
      summary: 探索列表
  /api/explore/receive:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 自增ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.ExploreReceiveResponse'
      summary: 探索领奖
  /api/explore/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 探索自增ID
        in: query
        name: id
        required: true
        type: integer
      - description: 英雄自增ID列表
        in: query
        items:
          type: integer
        name: heroIDs
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.ExploreSetResponse'
      summary: 探索设置
  /api/gm/addequip:
    post:
      consumes:
      - application/json
      parameters:
      - description: 装备ID
        in: query
        name: equipID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.GmCommonResponse'
      summary: 增加装备
  /api/gm/incrgold:
    post:
      consumes:
      - application/json
      parameters:
      - description: 金币数量
        in: query
        name: gold
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.GmCommonResponse'
      summary: 增加金币
...
swagger: "2.0"